# filebeat/input/filestream: Improve fingerprint performance