**PROPOSAL** (Please do not actively review until the proposal is accepted)

I recently stumbled upon a great blog post about the [filestream fingerprint](https://www.elastic.co/blog/introducing-filestream-fingerprint-mode) mode by @rdner, which caught my attention. After reading it, I noticed a few areas where the performance could be improved. For instance, the blog mentions the use of SHA256, a highly secure hashing algorithm. However, since we don't require such high levels of security for data hashing in this context, it would be better to explore faster hashing algorithms that can still ensure data integrity.

In the beats repository, **github.com/cespare/xxhash/v2** ([xxhash](https://github.com/Cyan4973/xxHash)) is used in multiple places which is a very fast cryptographically-insecure hash algorithm. But there's something even better we can use here i.e., xxh3.

See some benchmarks: https://xxhash.com

This is a [good blog](https://jolynch.github.io/posts/use_fast_data_algorithms/) comparing the different algorithms and advising what could be a better choice.

So, I used xxh3 (https://github.com/zeebo/xxh3) package used by apache/arrow, grafana, etc. I tested with the defaults on MacOS 13.6 (Apple M1 Max) with the already existing benchmark function (put `b.ResetTimer` to reset the timer after the creation of files because they are not necessary to the benchmark) and here are the results (it'd be better if the benchmarks were taken in a separate machine with lesser [noise](https://llvm.org/docs/Benchmarking.html) because my machine always has multiple instances of chrome running and few more things):

Used benchstat for comparing old and new benchmarks (count = 6)

```
                           │  bench.old  │             bench.new             │
                           │   sec/op    │   sec/op     vs base              │
GetFilesWithFingerprint-10   21.22m ± 2%   20.72m ± 1%  -2.37% (p=0.009 n=6)
```

```
                           │  bench.old   │              bench.new              │
                           │     B/op     │     B/op      vs base               │
GetFilesWithFingerprint-10   1.433Mi ± 0%   1.035Mi ± 0%  -27.78% (p=0.002 n=6)
```

```
                           │  bench.old  │             bench.new             │
                           │  allocs/op  │  allocs/op   vs base              │
GetFilesWithFingerprint-10   11.09k ± 0%   10.04k ± 0%  -9.48% (p=0.002 n=6)
```

In the benchmark, only 1000 are considered but in a production server, the number of files could be considerably much higher. So, the changes in this PR could be very beneficial when the user has opted for fingerprint mode. xxh3 also works well for smaller inputs and I assume the fingerprint length would mostly be small.

In case we plan to use an algorithm that works even better for very small inputs, [this](https://github.com/apache/arrow/blob/c5bce96ba626ad2558e10acddf8ef19ea06940e8/go/internal/hashing/hash_funcs.go#L58) may be beneficial to read.


Some more reading material, if interested:
- https://fastcompression.blogspot.com/2019/03/presenting-xxh3.html (xxh3)
- https://github.com/honeycombio/refinery/issues/749 (related to capacity of map)
- https://www.reddit.com/r/zfs/comments/wb50b7/will_zfs_support_the_hash_algorithm_xxhash_in_the/
- https://btrfs.readthedocs.io/en/latest/Checksumming.html#checksumming (btrfs' mention of xxhash as an option for checksum)
- https://fastcompression.blogspot.com/2019/03/presenting-xxh3.html#:~:text=Checksumming%20long%20inputs (xxhash popular among movie makers for file transfer verification)


## Proposed commit message

Improved the overall performance of the fingerprint mode by switching the hashing algorithm from SHA256 to XXH3. Also, removed an unnecessary check (not required for XXH3 anymore) and updated the tests. Also preallocating the maps (a better number could be decided) and some minor changes to the code.

## Checklist

- [x] My code follows the style guidelines of this project
- [x] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [ ] I have made corresponding change to the default configuration files
- [ ] I have added tests that prove my fix is effective or that my feature works
- [ ] I have added an entry in `CHANGELOG.next.asciidoc` or `CHANGELOG-developer.next.asciidoc`.

## How to test this PR locally

- Run go tests and benchmarks