@shmsr 2023-10-15
As this is a proposal, marking this PR as a draft. Also, I think as I updated the NOTICE.txt, it automatically tagged `elastic/beats-tech-leads` as a reviewer.